pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conversation-analysis-system"
version = "0.1.0"
description = "Automated conversation analysis pipeline for AI assistant quality insights"
requires-python = ">=3.11"
dependencies = [
    "pymongo>=4.6",
    "click>=8.1",
    "jinja2>=3.1",
    "textblob>=0.18",
    "sentence-transformers>=2.7",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.100",
    "pytest>=8.0",
    "mongomock>=4.1",
]

[project.scripts]
cas = "cas.cli:cli"

[tool.setuptools.packages.find]
include = ["cas*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
